Page components include as little logic as possible to make them more testable and re-usable. Fixes #1740
