+++ unlock-app/src/__tests__/components/content/DemoContent.test.js
@@ -1,4 +1,4 @@
-import { mapStateToProps } from '../../pages/demo'
+import { mapStateToProps } from '../../../components/content/DemoContent'
 
 describe('Demo', () => {
   describe('mapStateToProps', () => {
