--- unlock-app/src/pages/provider.js
@@ -1,76 +1,0 @@
-import PropTypes from 'prop-types'
-import React from 'react'
-import Link from 'next/link'
-import Head from 'next/head'
-import { connect } from 'react-redux'
-import Layout from '../components/interface/Layout'
-import UnlockPropTypes from '../propTypes'
-import { setProvider } from '../actions/provider'
-import withConfig from '../utils/withConfig'
-import { pageTitle } from '../constants'
-import BrowserOnly from '../components/helpers/BrowserOnly'
-
-export function Web3Provider({ setProvider, config, provider }) {
-  return (
-    <Layout title="Pick Web3 Provider">
-      <Head>
-        <title>{pageTitle('Pick Web3 Provider')}</title>
-      </Head>
-      <BrowserOnly>
-        <div>
-          <p>Pick web3 provider:</p>
-          <select
-            value={provider}
-            onChange={event => setProvider(event.target.value)}
-          >
-            {Object.keys(config.providers).map(name => {
-              return (
-                <option value={name} key={name}>
-                  {name}
-                </option>
-              )
-            })}
-          </select>
-        </div>
-        <p>
-          <Link href="/">
-            <a>Return</a>
-          </Link>
-        </p>
-      </BrowserOnly>
-    </Layout>
-  )
-}
-
-Web3Provider.displayName = 'Web3Provider'
-
-Web3Provider.propTypes = {
-  provider: UnlockPropTypes.provider.isRequired,
-  setProvider: PropTypes.func.isRequired,
-  config: UnlockPropTypes.configuration.isRequired,
-}
-
-const mapStateToProps = state => {
-  return {
-    provider: state.provider,
-  }
-}
-
-const mapDispatchToProps = dispatch => ({
-  setProvider: provider => dispatch(setProvider(provider)),
-})
-
-const Page = withConfig(
-  connect(
-    mapStateToProps,
-    mapDispatchToProps
-  )(Web3Provider)
-)
-
-const ProviderPage = pageProps => (
-  <BrowserOnly>
-    <Page {...pageProps} />
-  </BrowserOnly>
-)
-
-export default ProviderPage
