+++ unlock-app/src/pages/terms.js
@@ -1,680 +1,6 @@
 import React from 'react'
-import Head from 'next/head'
-import Link from 'next/link'
+import TermsContent from '../components/content/TermsContent'
 
-import Layout from '../components/interface/Layout'
-import Signature from '../components/interface/Signature'
-import {
-  Section,
-  Title,
-  SubTitle,
-  Paragraph,
-  UnorderedList,
-  ListItem,
-  Strong,
-} from '../components/Components'
-import { pageTitle } from '../constants'
-
-export const Terms = () => (
-  <Layout forContent>
-    <Head>
-      <title>{pageTitle('Terms of Service')}</title>
-    </Head>
-    <Section>
-      <Title>Terms of Service</Title>
-      <Paragraph>Last modified: January 22, 2019</Paragraph>
-      <Paragraph>
-        PLEASE READ THESE TERMS OF SERVICE (&ldquo;AGREEMENT&rdquo; OR
-        &ldquo;TERMS OF SERVICE&rdquo;) CAREFULLY BEFORE USING THE UNLOCK, INC.
-        SERVICES. It is a binding agreement between you or any user, including
-        users who are also contributors of content, (&ldquo;End User&rdquo; or
-        &ldquo;you&rdquo;) and Unlock, Inc. (&ldquo;Unlock&rdquo;). For purposes
-        of this Agreement, the Unlock &ldquo;Services&rdquo; include Unlock’s
-        website (Unlock-protocol.com), domain name, and its services, content,
-        applications and features, including any protocols and components
-        thereof, which may exist now or in the future and which it may offer
-        from time to time. Unlock’s &ldquo;Site&rdquo; means Unlock’s website
-        and its domain name.
-      </Paragraph>
-
-      <SubTitle>Acceptance of Terms</SubTitle>
-      <Paragraph>
-        The Services are offered on the condition that you accept this Agreement
-        without modification, including all other operating rules, policies and
-        procedures that may be published from time to time on or via the
-        Services.
-      </Paragraph>
-      <Paragraph>
-        BY CLICKING THE &ldquo;AGREE&rdquo; BUTTON, YOU (A) ACKNOWLEDGE THAT YOU
-        HAVE READ AND UNDERSTAND THIS AGREEMENT; (B) REPRESENT THAT YOU ARE 18
-        YEARS OF AGE OR OLDER; (C) ACCEPT THIS AGREEMENT AND AGREE THAT YOU ARE
-        LEGALLY BOUND BY ITS TERMS; AND (C) ALL REGISTRATION INFORMATION YOU
-        SUBMIT IS ACCURATE AND TRUTHFUL. IF YOU DO NOT AGREE TO THESE TERMS, DO
-        NOT ACCESS OR USE THE SERVICES.
-      </Paragraph>
-      <Paragraph>
-        In addition, some Services may be subject to additional terms and
-        conditions promulgated by Unlock from time to time; your use of such
-        Services is subject to those additional terms and conditions, which are
-        incorporated into this Agreement by this reference.
-      </Paragraph>
-      <Paragraph>
-        Please note that in all cases and to the maximum extent it is legally
-        permissible, Unlock may, in its sole discretion, refuse to offer the
-        Services (in whole or in part) to any person or entity and change its
-        eligibility criteria at any time.
-      </Paragraph>
-
-      <SubTitle>Changes and Updates</SubTitle>
-      <Paragraph>
-        Unlock reserves the right to modify or replace any of the terms of this
-        Agreement, or change, suspend, or discontinue the Services (including
-        without limitation, the availability of any feature, database, or
-        content) at any time by posting a notice on the Services or by sending
-        you an email. Unlock may also impose limits on certain Services or
-        restrict your access to parts or all of the Services without notice or
-        liability.
-      </Paragraph>
-      <Paragraph>
-        Please be aware that it is your responsibility to check this Agreement
-        periodically for changes. Your continued use of the Services following
-        the posting of any changes will mean you accept those changes.
-      </Paragraph>
-
-      <SubTitle>Privacy</SubTitle>
-      <Paragraph>
-        The privacy protocols and terms that apply to the Services are located
-        at{' '}
-        <Link href="/privacy">
-          <a>Privacy Policy</a>
-        </Link>{' '}
-        and is incorporated into this Agreement. For inquiries in regard to the
-        Privacy Policy, or to report a privacy related problem, please contact{' '}
-        <Link href="mailto:hello@unlock-protocol.com">
-          <a>hello@unlock-protocol.com</a>
-        </Link>
-        .
-      </Paragraph>
-
-      <SubTitle>Intellectual Property</SubTitle>
-      <Paragraph>
-        You hereby acknowledge that Unlock owns or has the license rights to all
-        right, title and interest in and to the Services, including without
-        limitation, the API, and all underlying software and technology and the
-        Intellectual Property Rights related thereto. &ldquo;Intellectual
-        Property Rights&rdquo; means any and all rights existing from time to
-        time under patent law, copyright law, trade secret law, trademark law,
-        unfair competition law, and any and all other proprietary rights, and
-        any and all applications, renewals, extensions and restorations thereof,
-        now or hereafter in force and effect worldwide.
-      </Paragraph>
-      <Paragraph>
-        Unless expressly agreed to subsequently in writing with Unlock, all
-        rights not expressly granted under this Agreement are hereby reserved.
-        Nothing in this Agreement gives you or anyone else a right to use any of
-        the trade names, trade marks, service marks, logos, domain names, and
-        other distinctive brand features related to Unlock Site or Services.
-      </Paragraph>
-
-      <SubTitle>Rules and Conduct</SubTitle>
-      <Paragraph>
-        As a condition of use and access to the Services, we expect that you
-        will not use the Services inappropriately. You promise not to use the
-        Services for any purpose that is prohibited by this Agreement or
-        applicable law or to intentionally damage, offend or injure anybody
-        else. You are responsible for all of your activity in connection with
-        the Services. For purposes of this Agreement, the term
-        &ldquo;Content&rdquo; includes, without limitation, any advertisements,
-        advice, suggestions, videos, audio clips, written forum comments,
-        information, data, text, photographs, software, scripts, graphics, and
-        interactive features generated, provided, or otherwise made accessible
-        by Unlock or its partners on or through the Services.
-      </Paragraph>
-      <Paragraph>
-        By way of example, and not as a limitation, you shall not (and shall not
-        permit any third party to) either (a) take any action, or (b) upload,
-        download, post, submit or otherwise distribute or facilitate
-        distribution of any material on or through the Services, that:
-      </Paragraph>
-      <UnorderedList>
-        <ListItem>
-          infringes any Intellectual Property Rights of any other person or
-          entity;
-        </ListItem>
-        <ListItem>
-          is unlawful, threatening, abusive, harassing, defamatory, libelous,
-          deceptive, fraudulent, invasive of another&rsquo;s privacy, tortious,
-          obscene, offensive, or profane;
-        </ListItem>
-        <ListItem>
-          constitutes unauthorized or unsolicited advertising, junk or bulk
-          e-mail (&ldquo;spamming&rdquo;);
-        </ListItem>
-        <ListItem>
-          contains software viruses or any other computer codes, files, or
-          programs that are designed or intended to disrupt, damage, limit or
-          interfere with the proper function of any software, hardware, or
-          telecommunications equipment or to damage or obtain unauthorized
-          access to any system, data, password or other information of Unlock or
-          any third party; or
-        </ListItem>
-        <ListItem>
-          impersonates any person or entity, including any employee or
-          representative of Unlock.
-        </ListItem>
-      </UnorderedList>
-      <Paragraph>
-        Additionally, you shall not: (i) take any action that imposes or may
-        impose (as determined by Unlock in its sole discretion) an unreasonable
-        or disproportionately large load on Unlock&rsquo;s (or its third party
-        providers&rsquo;) infrastructure; (ii) interfere or attempt to interfere
-        with the proper working of the Services or any activities conducted on
-        the Services; (iii) bypass any measures Unlock may use to prevent or
-        restrict access to the Services (or other accounts, computer systems or
-        networks connected to the Services); or (iv) use manual or automated
-        software, devices, or other processes to &ldquo;crawl&rdquo; or
-        &ldquo;spider&rdquo; any page of the Services.
-      </Paragraph>
-      <Paragraph>
-        You shall not (directly or indirectly): (i) decipher, decompile,
-        disassemble, reverse engineer or otherwise attempt to derive any source
-        code or underlying ideas or algorithms of any part of the Services,
-        except to the limited extent applicable laws specifically prohibit such
-        restriction, or (ii) modify any part of the Services. You shall abide by
-        all applicable local, state, national and international laws and
-        regulations.
-      </Paragraph>
-      <Paragraph>
-        Unlock reserves the right to remove any Content from the Services at any
-        time, for any reason (including, but not limited to, upon receipt of
-        claims or allegations from third parties or authorities relating to such
-        Content or if Unlock is concerned that you may have violated this
-        Agreement), or for no reason at all.
-      </Paragraph>
-
-      <SubTitle>Registration</SubTitle>
-      <Paragraph>
-        As a condition to using certain aspects of the Services, you will be
-        required to register with Unlock. You are solely responsible for
-        activity that occurs on or associated with your account and shall be
-        responsible for maintaining the confidentiality of your Unlock account
-        details, to the extent applicable. You shall never use another
-        user&rsquo;s account without such other user&rsquo;s express permission.
-        You will immediately notify Unlock in writing of any unauthorized use of
-        your account, or other account-related security breach of which you are
-        aware.
-      </Paragraph>
-
-      <SubTitle>Fees and Payment</SubTitle>
-      <Paragraph>
-        Unlock reserves the right to require payment of fees for the Services,
-        including with regard to any features at any time in the future, whether
-        or not Unlock require payment currently. Should you elect to subscribe
-        to a feature or component of the Services, you shall pay all applicable
-        fees, as described on the Site or Services. Unlock reserves the right to
-        change its prices and to institute new charges at any time, upon ten
-        (10) days prior notice to you, which may be sent by email or posted on
-        the Site or otherwise.
-      </Paragraph>
-      <Paragraph>
-        Please note that use of or access to the Services by you following such
-        notification constitutes your affirmative acceptance of any new or
-        increased charges.
-      </Paragraph>
-
-      <SubTitle>Third Party Sites</SubTitle>
-      <Paragraph>
-        The Services may permit you to link to other websites or resources on
-        the Internet, and other websites or resources may contain links to the
-        Site or the Services. When you access third party websites, services,
-        products or the data from them,{' '}
-        <Strong>you do so 100% at your own risk</Strong>. These other websites
-        are not under Unlock&rsquo;s control, and you hereby acknowledge that
-        Unlock is not responsible or liable for any aspects of such websites,
-        services, products or data, including but not limited to any content,
-        functions, accuracy, legality, appropriateness or any other aspects
-        whatsoever.
-      </Paragraph>
-      <Paragraph>
-        The inclusion of any links does not imply endorsement by Unlock or any
-        association with its operators. As a result, you hereby acknowledge and
-        agree that Unlock shall not be responsible or liable, directly or
-        indirectly, for any liability, damage or loss associated or in
-        connection with the access to, use of or reliance on any Content, or any
-        such non-Unlock sites, goods, services, technologies, or offerings.
-      </Paragraph>
-
-      <SubTitle>Blockchain Transactions And Third Parties</SubTitle>
-      <Paragraph>
-        Unlock reserves the right to refuse to process or to cancel any pending
-        Digital Currency Transaction as required by law or in response to a
-        subpoena, court order, or other binding government order or to enforce
-        transaction limits. Unlock cannot reverse a Digital Currency Transaction
-        which has been broadcast to a blockchain. Know that the use of or access
-        to any facet or component of a blockchain is solely and 100% at your own
-        risk.
-      </Paragraph>
-      <Paragraph>
-        Be aware that Unlock has no control over, liability for, or the
-        delivery, quality, safety, legality or any other aspect of anything that
-        you may purchase or sell to or from a third party (including other users
-        of the Services). Unlock is not responsible for ensuring that a buyer or
-        a seller you may transact with will actually complete the transaction or
-        is authorized to do so. If you experience a problem with any purchase or
-        sale or otherwise in connection with using the Services, or if you have
-        a dispute with such third party, you must resolve the dispute directly
-        with that third party. If you believe a third party has behaved in a
-        fraudulent, misleading, or inappropriate manner, or if you cannot
-        adequately resolve a dispute with a third party, you may notify Unlock
-        Support at{' '}
-        <Link href="mailto:unlock-protocol.com">
-          <a>hello@unlock-protocol.com</a>
-        </Link>{' '}
-        so that we may consider what action to take, if any, at our sole
-        discretion. Your use or access to the Services, and any interaction with
-        any third party services, products or protocols, especially including
-        any blockchain technology, is 100% at your own risk.
-      </Paragraph>
-
-      <SubTitle>Unlock and Site Content</SubTitle>
-      <Paragraph>
-        You agree that the Services may contain Content provided by Unlock or
-        other third parties and that such Content is protected by copyrights,
-        trademarks, service marks, patents, trade secrets or other proprietary
-        rights and laws. You shall abide by all copyright notices, information,
-        and restrictions contained in any Content accessed or used via the
-        Services. You shall not sell, license, rent, modify, distribute, copy,
-        reproduce, transmit, publicly display, publicly perform, publish, adapt,
-        edit, create derivative works from, or otherwise exploit any Content or
-        third party submissions or other proprietary rights not owned by you and
-        in any way that violates any third party right.
-      </Paragraph>
-
-      <SubTitle>Termination</SubTitle>
-      <Paragraph>
-        Unlock may terminate your access to all or any part of the Services at
-        any time, with or without cause, with or without notice, effective
-        immediately, which may result in the forfeiture and destruction of all
-        information associated with your account or use of the Services. If you
-        wish to terminate your account, you may do so by following the
-        instructions on the Site or as otherwise provided via the Services. Any
-        fees paid hereunder are non-refundable unless otherwise expressly set
-        forth herein. All provisions of this Agreement which by their nature
-        should survive termination shall survive termination, including, without
-        limitation, ownership provisions, warranty disclaimers, indemnity and
-        limitations of liability.
-      </Paragraph>
-      <Paragraph>
-        Upon termination: (i)all rights granted to you under this Agreement will
-        also terminate; and (ii)you must cease all use of the Services and any
-        Content that does not belong to you, and delete all copies of such from
-        your equipment and account. Termination will not limit any of Unlocks
-        rights or remedies at law or in equity.
-      </Paragraph>
-
-      <SubTitle>Warranty Disclaimer</SubTitle>
-      <Paragraph>
-        Unlock has no special relationship with or fiduciary duty to you. You
-        release Unlock, our directors, employees, agents and successors from
-        claims, demands any and all losses, damages, rights, claims, and actions
-        of any kind from all liability for you having acquired or not acquired
-        Content through the Site. If you are a California resident, you hereby
-        waive California Civil Code Section 1542, which states:{' '}
-        <Strong>
-          &ldquo;A general release does not extend to claims which the creditor
-          does not know or suspect to exist in his favor at the time of
-          executing the release, which, if known by him must have materially
-          affected his settlement with the debtor.&rdquo;
-        </Strong>
-      </Paragraph>
-      <Paragraph>
-        The Services may contain, or direct you to websites containing,
-        information that some people may find offensive or inappropriate. Unlock
-        makes no representations concerning any content contained in or accessed
-        through the Site, and Unlock will not be responsible or liable for the
-        accuracy, copyright compliance, legality or decency of material
-        contained in or accessed through the Services.
-      </Paragraph>
-      <Paragraph>
-        FURTHERMORE, ESPECIALLY CONSIDERING THE SERVICES ARE PROVIDED WITHOUT
-        CHARGE, THE SERVICES (INCLUDING, WITHOUT LIMITATION, ANY CONTENT AND THE
-        OFFERINGS ASSOCIATED WITH THE SITES) ARE PROVIDED{' '}
-        <Strong>
-          &ldquo;AS IS&rdquo; AND &ldquo;AS AVAILABLE&rdquo; AND ARE WITHOUT
-          WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING, BUT NOT LIMITED
-          TO, THE IMPLIED WARRANTIES OF TITLE, NON- INFRINGEMENT,
-          MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE, AND ANY
-          WARRANTIES IMPLIED BY ANY COURSE OF PERFORMANCE OR USAGE OF TRADE, ALL
-          OF WHICH ARE EXPRESSLY DISCLAIMED.
-        </Strong>{' '}
-        UNLOCK, AND ITS DIRECTORS, EMPLOYEES, AGENTS, SUPPLIERS, PARTNERS AND
-        CONTENT PROVIDERS DO NOT WARRANT THAT: (A) THE SERVICES WILL BE SECURE
-        OR AVAILABLE AT ANY PARTICULAR TIME OR LOCATION; (B) ANY DEFECTS OR
-        ERRORS WILL BE CORRECTED; (C) ANY CONTENT OR SOFTWARE AVAILABLE AT OR
-        THROUGH THE SERVICES IS FREE OF VIRUSES OR OTHER HARMFUL COMPONENTS; OR
-        (D) THE RESULTS OF USING THE SERVICES WILL MEET YOUR REQUIREMENTS.{' '}
-        <Strong>
-          YOUR USE OF AND ACCESS TO THE SERVICES IS SOLELY AT YOUR OWN RISK.
-        </Strong>
-      </Paragraph>
-      <Paragraph>
-        SOME STATES DO NOT ALLOW LIMITATIONS ON HOW LONG AN IMPLIED WARRANTY
-        LASTS, SO THE ABOVE LIMITATIONS MAY NOT APPLY TO YOU BUT YOU AGREE THAT
-        THE LIMITATIONS AND DISCLAIMERS WILL APPLY TO THE MAXIMUM EXTENT
-        PERMITTED UNDER APPLICABLE LAW.
-      </Paragraph>
-      <Paragraph>
-        Electronic Communications Privacy Act Notice (18 USC 2701-2711): UNLOCK
-        MAKES NO GUARANTEE OF CONFIDENTIALITY OR PRIVACY OF ANY COMMUNICATION OR
-        INFORMATION TRANSMITTED ON THE SITES OR ANY WEBSITE LINKED TO THE
-        SERVICES. Unlock will not be liable for the privacy of email addresses,
-        registration and identification information, disk space, communications,
-        confidential or trade-secret information, or any other Content stored on
-        or associated with Unlock&rsquo;s Services, equipment, or any
-        information transmitted over networks accessed by the Site, or otherwise
-        connected with your use of the Services.
-      </Paragraph>
-
-      <SubTitle>Indemnification</SubTitle>
-      <Paragraph>
-        You hereby agree that you shall defend, indemnify, and hold harmless
-        Unlock, its affiliates and each of its, and its affiliates employees,
-        contractors, directors, suppliers and representatives from all
-        liabilities, claims, and expenses, including reasonable attorneys&rsquo;
-        fees, that arise from or relate to (i) your use or misuse of, or access
-        to, the Site, Services, or Content, (ii) your violation of this
-        Agreement, or (iii) infringement by you, or any third party using the
-        your account, of any intellectual property or other right of any person
-        or entity. Unlock reserves the right to assume the exclusive defense and
-        control of any matter otherwise subject to indemnification by you, in
-        which event you will assist and cooperate with Unlock in asserting any
-        available defenses.
-      </Paragraph>
-
-      <SubTitle>Limitation of Liability</SubTitle>
-      <Paragraph>
-        IN NO EVENT SHALL UNLOCK, NOR ITS DIRECTORS, EMPLOYEES, AGENTS,
-        PARTNERS, SUPPLIERS OR CONTENT PROVIDERS, BE LIABLE UNDER CONTRACT,
-        TORT, STRICT LIABILITY, NEGLIGENCE OR ANY OTHER LEGAL OR EQUITABLE
-        THEORY WITH RESPECT TO THE SERVICES (INCLUDING, WITHOUT LIMITATION, ANY
-        CONTENT OR ANY ASPECTS ASSOCIATED WITH THE SITES) (I) FOR ANY LOST
-        PROFITS, DATA LOSS, COST OF PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES,
-        OR SPECIAL, INDIRECT, INCIDENTAL, PUNITIVE, OR CONSEQUENTIAL DAMAGES OF
-        ANY KIND WHATSOEVER, SUBSTITUTE GOODS OR SERVICES (HOWEVER ARISING),
-        (II) FOR ANY BUGS, VIRUSES, TROJAN HORSES, OR THE LIKE (REGARDLESS OF
-        THE SOURCE OF ORIGINATION), (III) FOR YOUR RELIANCE ON THE SERVICES OR
-        (IV) FOR ANY DIRECT DAMAGES IN EXCESS OF (IN THE AGGREGATE) THE GREATER
-        OF (A) THE FEES PAID BY YOU TO UNLOCK HEREUNDER WITHIN THE 12 MONTHS
-        PRIOR TO ACCRUAL OF THE CLAIM AND (B) ONE-HUNDRED U.S. DOLLARS
-        ($100.00).{' '}
-        <Strong>
-          YOU HEREBY AGREE THAT YOU FULLY UNDERSTAND THE IMPLICATIONS OF THESE
-          CLAUSES (WHICH LIMIT UNLOCK’S LIABILITY AND YOUR REMEDIES) AND THAT
-          THE EXCLUSIONS AND LIMITATION OF LIABILITY SHALL APPLY TO THE MAXIMUM
-          EXTENT PERMITTED UNDER APPLICABLE LAW.
-        </Strong>
-      </Paragraph>
-
-      <SubTitle>Compliance with Local Laws</SubTitle>
-      <Paragraph>
-        Unlock makes no representation that accessing or using the Services is
-        not prohibited where you are located. The Content and Services are based
-        in the state of New York, in the United States and provided for access
-        and use only by persons located in the United States. You acknowledge
-        that you may not be able to access all or some of the Content and
-        Services outside of the United States and that access thereto may not be
-        legal. By accessing the Services or Content, you do so at your own
-        initiative and risk and you are responsible for compliance with all
-        local laws.
-      </Paragraph>
-
-      <SubTitle>Export Regulation</SubTitle>
-      <Paragraph>
-        The Services and any components thereof may be subject to US export
-        control laws, including the US Export Administration Act and its
-        associated regulations. You shall not, directly or indirectly, export,
-        re-export, or release the Services or any components thereof, or make
-        the Services or any components thereof accessible from, any jurisdiction
-        or country to which export, re-export, or release is prohibited by law,
-        rule, or regulation. You shall comply with all applicable federal laws,
-        regulations, and rules, and complete all required undertakings
-        (including obtaining any necessary export license or other governmental
-        approval), prior to exporting, re-exporting, releasing, or otherwise
-        making the Services or any component thereof available outside the US.
-      </Paragraph>
-
-      <SubTitle>Dispute Resolution</SubTitle>
-      <Paragraph>
-        A printed version of the Agreement and of any notice given in electronic
-        form shall be admissible in judicial or administrative proceedings based
-        upon or relating to this Agreement to the same extent and subject to the
-        same conditions as other business documents and records originally
-        generated and maintained in printed form. You and Unlock agree that any
-        cause of action arising out of or related to the Site, Services or any
-        Content must commence within one (1) year after the cause of action
-        arose; otherwise, such cause of action is permanently barred.
-      </Paragraph>
-      <Paragraph>
-        This Agreement shall be governed by and construed in accordance with the
-        laws of the State of New York, excluding its conflicts of law rules, and
-        the United States of America. Any dispute arising from or relating to
-        the subject matter of this Agreement shall be finally settled by
-        arbitration in New York City, New York, using the English language in
-        accordance with the Arbitration Rules and Procedures of JAMS then in
-        effect, by one commercial arbitrator with substantial experience in
-        resolving intellectual property and commercial contract disputes, who
-        shall be selected from the appropriate list of JAMS arbitrators in
-        accordance with the Arbitration Rules and Procedures of JAMS. The
-        prevailing party in the arbitration shall be entitled to receive
-        reimbursement of its reasonable expenses (including reasonable
-        attorneys&rsquo; fees, expert witness fees and all other expenses)
-        incurred in connection therewith. Judgment upon the award so rendered
-        may be entered in a court having jurisdiction or application may be made
-        to such court for judicial acceptance of any award and an order of
-        enforcement, as the case may be. Notwithstanding the foregoing, each
-        party shall have the right to institute an action in a court of proper
-        jurisdiction for injunctive or other equitable relief pending a final
-        decision by the arbitrator. For all purposes of this Agreement, the
-        parties consent to exclusive jurisdiction and venue in the United States
-        Federal Courts located in the Northern District of New York. Use of the
-        Services is not authorized in any jurisdiction that does not give effect
-        to all provisions of the Agreement, including without limitation, this
-        section.
-      </Paragraph>
-
-      <SubTitle>Integration and Severability</SubTitle>
-      <Paragraph>
-        The Agreement is the entire agreement between you and Unlock with
-        respect to the Services and use of the Site, and supersede all prior or
-        contemporaneous communications and proposals (whether oral, written or
-        electronic) between you and Unlock with respect to the Site. If any
-        provision of the Agreement is found to be unenforceable or invalid, that
-        provision will be limited or eliminated to the minimum extent necessary
-        so that the Agreement will otherwise remain in full force and effect and
-        enforceable. The failure of either party to exercise in any respect any
-        right provided for herein shall not be deemed a waiver of any further
-        rights hereunder.
-      </Paragraph>
-
-      <SubTitle>Miscellaneous</SubTitle>
-      <Paragraph>
-        Unlock shall not be liable for any failure to perform its obligations
-        hereunder where such failure results from any cause beyond
-        Unlock&rsquo;s reasonable control, including, without limitation,
-        mechanical, electronic or communications failure or degradation
-        (including &ldquo;line-noise&rdquo; interference). The Agreement are
-        personal to you, and are not assignable, transferable or sublicensable
-        by you except with Unlock&rsquo;s prior written consent. Unlock may
-        assign, transfer or delegate any of its rights and obligations hereunder
-        without consent. No agency, partnership, joint venture, or employment
-        relationship is created as a result of the Agreement and neither party
-        has any authority of any kind to bind the other in any respect. In any
-        action or proceeding to enforce rights under the Agreement, the
-        prevailing party, and if there isn’t one, then the substantially
-        prevailing party, will be entitled to recover costs and attorneys&rsquo;
-        fees. All notices under the Agreement will be in writing and will be
-        deemed to have been duly given when received, if personally delivered or
-        sent by certified or registered mail, return receipt requested; when
-        receipt is electronically confirmed, if transmitted by e-mail; or the
-        day after it is sent, if sent for next day delivery by recognized
-        overnight delivery service.
-      </Paragraph>
-
-      <SubTitle>Trademark Notices</SubTitle>
-      <Paragraph>
-        Unless otherwise indicated, the Agreement and all Content provided in
-        connection with the Site or Services are copyright Unlock, Inc, all
-        rights reserved.
-      </Paragraph>
-      <Paragraph>
-        The names of actual companies and products mentioned at the Site may be
-        the trademarks of their respective owners.
-      </Paragraph>
-      <Paragraph>
-        <Strong>A. Procedure for Reporting Copyright Infringements:</Strong>
-      </Paragraph>
-      <Paragraph>
-        If you believe that Content residing on or accessible through the Unlock
-        web site or service infringes a copyright, please send a notice of
-        copyright infringement containing the following information to the
-        Designated Agent listed below:
-      </Paragraph>
-      <UnorderedList>
-        <ListItem>
-          A physical or electronic signature of a person authorized to act on
-          behalf of the owner of the copyright that has been allegedly
-          infringed;
-        </ListItem>
-        <ListItem>
-          Identification of works or materials being infringed;
-        </ListItem>
-        <ListItem>
-          Identification of the Content that is claimed to be infringing
-          including information regarding the location of the Content that the
-          copyright owner seeks to have removed, with sufficient detail so that
-          Unlock is capable of finding and verifying its existence;
-        </ListItem>
-        <ListItem>
-          Contact information about the notifier including address, telephone
-          number and, if available, e-mail address;
-        </ListItem>
-        <ListItem>
-          A statement that the notifier has a good faith belief that the Content
-          is not authorized by the copyright owner, its agent, or the law; and
-        </ListItem>
-        <ListItem>
-          A statement made under penalty of perjury that the information
-          provided is accurate and the notifying party is authorized to make the
-          complaint on behalf of the copyright owner.
-        </ListItem>
-      </UnorderedList>
-      <Paragraph>
-        <Strong>
-          B. Once Proper Bona Fide Infringement Notification is Received by the
-          Designated Agent:
-        </Strong>
-      </Paragraph>
-      <Paragraph>It is Unlock&rsquo;s policy:</Paragraph>
-      <UnorderedList>
-        <ListItem>
-          to remove or disable access to the infringing Content;
-        </ListItem>
-        <ListItem>
-          to notify the Content provider, member or user that it has removed or
-          disabled access to the Content; and
-        </ListItem>
-        <ListItem>
-          that repeat offenders will have the infringing Content removed from
-          the system and that Unlock will terminate such content
-          provider&rsquo;s, member&rsquo;s or user&rsquo;s access to the
-          service.
-        </ListItem>
-      </UnorderedList>
-      <Paragraph>
-        <Strong>
-          C. Procedure to Supply a Counter-Notice to the Designated Agent:
-        </Strong>
-      </Paragraph>
-      <Paragraph>
-        If the Content provider, member or user believes that the Content that
-        was removed or to which access was disabled is either not infringing, or
-        the Content provider, member or user believes that it has the right to
-        post and use such Content from the copyright owner, the copyright
-        owner&rsquo;s agent, or pursuant to the law, the content provider,
-        member or user must send a counter-notice containing the following
-        information to the Designated Agent listed below:
-      </Paragraph>
-      <UnorderedList>
-        <ListItem>
-          A physical or electronic signature of the Content provider, member or
-          user;
-        </ListItem>
-        <ListItem>
-          Identification of the Content that has been removed or to which access
-          has been disabled and the location at which the Content appeared
-          before it was removed or disabled;
-        </ListItem>
-        <ListItem>
-          A statement that the Content provider, member or user has a good faith
-          belief that the Content was removed or disabled as a result of mistake
-          or a misidentification of the Content; and
-        </ListItem>
-        <ListItem>
-          Content provider&rsquo;s, member&rsquo;s or user&rsquo;s name,
-          address, telephone number, and, if available, e-mail address and a
-          statement that such person or entity consents to the jurisdiction of
-          the Federal Court for the judicial district in which the content
-          provider&rsquo;s, member&rsquo;s or user&rsquo;s address is located,
-          or if the Content provider&rsquo;s, member&rsquo;s or user&rsquo;s
-          address is located outside the United States, for any judicial
-          district in which Unlock is located, and that such person or entity
-          will accept service of process from the person who provided
-          notification of the alleged infringement.
-        </ListItem>
-      </UnorderedList>
-      <Paragraph>
-        If a counter-notice is received by the Designated Agent, Unlock may send
-        a copy of the counter-notice to the original complaining party informing
-        that person that it may replace the removed Content or cease disabling
-        it in 10 business days. Unless the copyright owner files an action
-        seeking a court order against the Content provider, member or user, the
-        removed Content may be replaced, or access to it restored, in 10 to 14
-        business days or more after receipt of the counter-notice, at
-        Unlock&rsquo;s sole discretion.
-      </Paragraph>
-      <Paragraph>
-        Please contact Unlock&rsquo;s Designated Agent to Receive Notification
-        of Claimed Infringement at the following address:
-      </Paragraph>
-      <Paragraph>
-        Unlock Inc,
-        <br />
-        163 Wyckoff Street,
-        <br />
-        Brooklyn, NY, 11217
-        <br />
-        <Link href="mailto:hello@unlock-protocol.com">
-          <a>hello@unlock-protocol.com</a>
-        </Link>
-      </Paragraph>
-
-      <SubTitle>Contact</SubTitle>
-      <Paragraph>You may contact Unlock at the following address:</Paragraph>
-      <Paragraph>
-        Unlock Inc,
-        <br />
-        163 Wyckoff Street,
-        <br />
-        Brooklyn, NY, 11217
-        <br />
-        <Link href="mailto:hello@unlock-protocol.com">
-          <a>hello@unlock-protocol.com</a>
-        </Link>
-      </Paragraph>
-    </Section>
-
-    <Signature />
-  </Layout>
-)
+const Terms = () => <TermsContent />
 
 export default Terms
